requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.5.0
django-filter==24.3
python-decouple==3.8
python-dotenv==1.0.1
psycopg2==2.9.10
drf-yasg==1.21.9
Pillow==10.4.0
